+++ example/lib/package_structure/package_tools/tool_body_one.dart
@@ -14,6 +14,7 @@ class _ToolBodyOneState extends State<ToolBodyOne> {
     return MihPackageToolBody(
       backgroundColor: MihColors.primary(),
       bodyItem: MihSingleChildScroll(
+        scrollbarOn: true,
         child: Column(
           mainAxisAlignment: MainAxisAlignment.start,
           children: [